src/requirements.txt
streamlit
requests
praw
tweepy
google.generativeai
requests
streamlit
